requirements.txt
async_timeout==3.0.1
discord==1.7.3
python-dotenv==0.18.0
